backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
websockets>=12.0
google-genai>=0.2.0
# For the ADK: wait for exact pip package, might be google-adk
google-adk>=0.5.0
google-cloud-firestore>=2.16.0
google-cloud-aiplatform>=1.60.0
python-dotenv>=1.0.1
pydantic>=2.7.0
pydantic-settings>=2.2.1
openai>=1.0.0
